% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_module.R
\name{filter_module_server}
\alias{filter_module_server}
\title{Create filter module server logic}
\usage{
filter_module_server(id, metadata, distinct_values, initial_value = NULL)
}
\arguments{
\item{id}{Character. The module ID}

\item{metadata}{Metadata for the column}

\item{distinct_values}{Distinct values for categorical columns}

\item{initial_value}{Initial filter value}
}
\value{
List of reactive values
}
\description{
Create filter module server logic
}
