% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_module.R
\name{filter_module_ui}
\alias{filter_module_ui}
\title{Create filter module UI components}
\usage{
filter_module_ui(id, column_info, initial_value = NULL)
}
\arguments{
\item{id}{Character. The module ID}

\item{column_info}{List of metadata and distinct values dataframes}

\item{initial_value}{Vector. Initial filter value(s)}
}
\value{
Shiny UI element
}
\description{
Create filter module UI components
}
