% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_helpers.R
\name{get_demo_column_info_path}
\alias{get_demo_column_info_path}
\title{Get path to demo column info directory}
\usage{
get_demo_column_info_path()
}
\value{
Character path to the demo column info directory
}
\description{
Get path to demo column info directory
}
\examples{
column_info_dir <- get_demo_column_info_path()
}
