% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_helpers.R
\name{get_demo_database_path}
\alias{get_demo_database_path}
\title{Get path to demo database}
\usage{
get_demo_database_path()
}
\value{
Character path to the demo DuckDB database
}
\description{
Get path to demo database
}
\examples{
db_path <- get_demo_database_path()
}
