% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_helpers.R
\name{get_demo_pool}
\alias{get_demo_pool}
\title{Create demo database pool}
\usage{
get_demo_pool()
}
\value{
A pool connection to the demo database
}
\description{
Create demo database pool
}
\examples{
if (interactive()) {
  pool <- get_demo_pool()
  DBI::dbListTables(pool)
  pool::poolClose(pool)
}
}
