% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_helpers.R
\name{get_demo_storage_info}
\alias{get_demo_storage_info}
\title{Get demo storage info}
\usage{
get_demo_storage_info()
}
\value{
List with storage configuration for demo data
}
\description{
Get demo storage info
}
\examples{
storage_info <- get_demo_storage_info()
}
