% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_column_info.R
\name{get_storage_location}
\alias{get_storage_location}
\title{Get storage location helper function}
\usage{
get_storage_location(
  storage_type,
  column_info_dir,
  adls_endpoint,
  adls_container,
  sas_token
)
}
\arguments{
\item{storage_type}{Either "local" or "adls"}

\item{column_info_dir}{Path for local storage}

\item{adls_endpoint}{ADLS endpoint URL}

\item{adls_container}{ADLS container name}

\item{sas_token}{ADLS SAS token}
}
\value{
List with storage information
}
\description{
Get storage location helper function
}
