% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_builder.R
\name{plot_builder_server}
\alias{plot_builder_server}
\title{Create plot builder server logic}
\usage{
plot_builder_server(id, fetched_data)
}
\arguments{
\item{id}{Character. The module ID}

\item{fetched_data}{Reactive. Data from data_fetcher}
}
\value{
List of reactive expressions
}
\description{
Create plot builder server logic
}
