% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_builder.R
\name{plot_builder_ui}
\alias{plot_builder_ui}
\title{Create plot builder UI components}
\usage{
plot_builder_ui(id)
}
\arguments{
\item{id}{Character. The module ID}
}
\value{
A list of Shiny UI elements
}
\description{
Create plot builder UI components
}
