% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Database Analysis Application}
\usage{
run_app(pool, column_info_dir = tempdir())
}
\arguments{
\item{pool}{A database connection pool object}

\item{column_info_dir}{Directory containing column information files}
}
\description{
Run the Shiny Database Analysis Application
}
