% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydbanalysis-package.R
\docType{package}
\name{shinydbanalysis-package}
\alias{shinydbanalysis}
\alias{shinydbanalysis-package}
\title{shinydbanalysis: What the Package Does (One Line, Title Case)}
\description{
A Shiny-based tool for filtering, grouping, and summarizing database tables.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
