% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_picker.R
\name{table_picker_ui}
\alias{table_picker_ui}
\title{Create table picker UI components}
\usage{
table_picker_ui(id)
}
\arguments{
\item{id}{Character. The module ID}
}
\value{
A Shiny UI element
}
\description{
Create table picker UI components
}
