% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_info_generator.R
\name{create_column_info}
\alias{create_column_info}
\title{Generate column information for a table}
\usage{
create_column_info(tablename, pool, output_dir = "column_info")
}
\arguments{
\item{tablename}{Character. Name of the table to analyze}

\item{pool}{Pool object. Database connection pool}

\item{output_dir}{Character. Directory to save column info files. Defaults to "column_info"}
}
\value{
List of column information (invisibly)
}
\description{
Generate column information for a table
}
