% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fetcher.R
\name{data_fetcher_server}
\alias{data_fetcher_server}
\title{Create data fetcher server logic}
\usage{
data_fetcher_server(
  id,
  pool,
  selected_table_name,
  selected_tbl_ref_without_restricted_columns,
  where_clause,
  needs_summary,
  group_vars,
  summary_specs
)
}
\arguments{
\item{id}{Character. The module ID}

\item{pool}{Database connection pool}

\item{table_builder}{Table picker module instance}

\item{filter_builder}{Filter builder module instance}

\item{summary_builder}{Summary builder module instance}
}
\value{
List of reactive expressions
}
\description{
Create data fetcher server logic
}
