% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fetcher.R
\name{data_fetcher_server}
\alias{data_fetcher_server}
\title{Create data fetcher server logic}
\usage{
data_fetcher_server(
  id,
  pool,
  table_info,
  filter_builder,
  group_builder,
  summary_builder
)
}
\arguments{
\item{id}{Character. The module ID}

\item{pool}{Database connection pool}

\item{table_info}{Table picker module instance}

\item{filter_builder}{Filter builder module instance}

\item{group_builder}{Group builder module instance}

\item{summary_builder}{Summary builder module instance}
}
\value{
List of reactive expressions
}
\description{
Create data fetcher server logic
}
