% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fetcher.R
\name{data_fetcher_ui}
\alias{data_fetcher_ui}
\title{Create data fetcher UI components}
\usage{
data_fetcher_ui(id, style = "hover")
}
\arguments{
\item{id}{Character. The module ID}

\item{style}{Character. Either "collapsible" or "hover" for the preview display style}
}
\value{
A Shiny UI element
}
\description{
Create data fetcher UI components
}
