% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_shinydbanalysis_app.R
\name{demo_shinydbanalysis_app}
\alias{demo_shinydbanalysis_app}
\title{Run the Shiny Database Analysis Application}
\usage{
demo_shinydbanalysis_app(
  pool,
  storage_type = "local",
  column_info_dir = "column_info",
  adls_endpoint = NULL,
  adls_container = NULL,
  sas_token = NULL,
  restricted_columns = character(0)
)
}
\arguments{
\item{pool}{A database connection pool object}

\item{storage_type}{Either "local" or "adls"}

\item{column_info_dir}{Path for storing column info parquet files on either local or adls container (default: "column_info")}

\item{adls_endpoint}{ADLS endpoint URL (required if storage_type = "adls")}

\item{adls_container}{ADLS container name (required if storage_type = "adls")}

\item{sas_token}{ADLS SAS token (required if storage_type = "adls")}
}
\description{
Run the Shiny Database Analysis Application
}
