% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_builder.R
\name{filter_builder_server}
\alias{filter_builder_server}
\title{Create filter builder server logic}
\usage{
filter_builder_server(id, storage_info, selected_table, restricted_columns)
}
\arguments{
\item{id}{Character. The module ID}

\item{storage_info}{List with storage configuration}

\item{selected_table}{Reactive. Selected table name}

\item{restricted_columns}{Reactive. Columns to restrict}
}
\value{
List of reactive expressions
}
\description{
Create filter builder server logic
}
