% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_builder.R
\name{filter_builder_ui}
\alias{filter_builder_ui}
\title{Create filter builder UI components}
\usage{
filter_builder_ui(id)
}
\arguments{
\item{id}{Character. The module ID}
}
\value{
A list of Shiny UI elements
}
\description{
Create filter builder UI components
}
