% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_module.R
\name{filter_module_server}
\alias{filter_module_server}
\title{Create filter module server}
\usage{
filter_module_server(id, column_info, initial_value = NULL)
}
\arguments{
\item{id}{Character. The module ID}

\item{column_info}{List. Column information}

\item{initial_value}{Vector. Initial filter value(s)}
}
\value{
List of reactive expressions
}
\description{
Create filter module server
}
