% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_module.R
\name{filter_module_ui}
\alias{filter_module_ui}
\title{Individual Filter Module with Improved Layout and Date Support
Create filter module UI}
\usage{
filter_module_ui(id, column_info, initial_value = NULL)
}
\arguments{
\item{id}{Character. The module ID}

\item{column_info}{List. Column information containing name, type, and possible values}

\item{initial_value}{Vector. Initial filter value(s)}
}
\value{
A Shiny UI element
}
\description{
Individual Filter Module with Improved Layout and Date Support
Create filter module UI
}
