% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_builder.R
\name{group_builder_server}
\alias{group_builder_server}
\title{Create group builder server}
\usage{
group_builder_server(id, selected_table, column_info)
}
\arguments{
\item{id}{Character. The module ID}

\item{selected_table}{Reactive. Selected table from table_picker}

\item{column_info}{Reactive. Column info list containing metadata and distinct values}
}
\value{
List of reactive expressions
}
\description{
Create group builder server
}
