% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_builder.R
\name{group_builder_ui}
\alias{group_builder_ui}
\title{Create group builder UI components}
\usage{
group_builder_ui(id)
}
\arguments{
\item{id}{Character. The module ID}
}
\value{
A list of Shiny UI elements
}
\description{
Create group builder UI components
}
