% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny application with debugging}
\usage{
run_app(pool, column_info_dir = "column_info")
}
\arguments{
\item{pool}{Database connection pool}

\item{column_info_dir}{Path to directory containing column info files}
}
\description{
Run the Shiny application with debugging
}
