% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydbanalysis-package.R
\docType{package}
\name{shinydbanalysis-package}
\alias{shinydbanalysis}
\alias{shinydbanalysis-package}
\title{shinydbanalysis: A database-agnostic Shiny-based tool for filtering, grouping, and summarizing database tables.}
\description{
A Shiny-based tool for filtering, grouping, and summarizing database tables.
}
\author{
\strong{Maintainer}: Simon Coulombe \email{first.last@example.com}

}
\keyword{internal}
