% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_filter.R
\name{single_filter_demo}
\alias{single_filter_demo}
\title{Demo app for single_filter}
\usage{
single_filter_demo()
}
\value{
A Shiny app object
}
\description{
Demo app for single_filter
}
\examples{
if (interactive()) {
  single_filter_demo()
}
}
