% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_filter.R
\name{single_filter_server}
\alias{single_filter_server}
\title{Create single filter server logic}
\usage{
single_filter_server(id, single_column_info, initial_value = NULL)
}
\arguments{
\item{id}{Character. The module ID}

\item{single_column_info}{List with metadata and distinct_values}

\item{initial_value}{Initial filter value}
}
\value{
List of reactive values
}
\description{
Create single filter server logic
}
