% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_filter.R
\name{single_filter_ui}
\alias{single_filter_ui}
\title{Create single filter UI components}
\usage{
single_filter_ui(id, single_column_info, initial_value = NULL)
}
\arguments{
\item{id}{Character. The module ID}

\item{single_column_info}{List with metadata and distinct_values dataframes}

\item{initial_value}{Vector. Initial filter value(s)}
}
\value{
Shiny UI element
}
\description{
Create single filter UI components
}
