% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_builder.R
\name{summary_builder_server}
\alias{summary_builder_server}
\title{Create summary builder server}
\usage{
summary_builder_server(id, selected_table, column_info)
}
\arguments{
\item{id}{Character. The module ID}

\item{selected_table}{Reactive. Selected table from table_picker}

\item{column_info}{Reactive. Column info list containing metadata and distinct values}
}
\value{
List of reactive expressions containing summary specifications
}
\description{
Create summary builder server
}
