% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_builder.R
\name{summary_builder_ui}
\alias{summary_builder_ui}
\title{Create summary builder UI components}
\usage{
summary_builder_ui(id)
}
\arguments{
\item{id}{Character. The module ID}
}
\value{
A list of Shiny UI elements
}
\description{
Create summary builder UI components
}
