% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_picker.R
\name{table_picker_server}
\alias{table_picker_server}
\title{Create table picker server}
\usage{
table_picker_server(id, pool, storage_info)
}
\arguments{
\item{id}{Character. The module ID}

\item{pool}{Pool object. Database connection pool}

\item{storage_info}{List containing storage configuration}
}
\value{
List of reactive expressions
}
\description{
Create table picker server
}
