% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_picker.R
\name{table_picker_server}
\alias{table_picker_server}
\title{Create table picker server}
\usage{
table_picker_server(id, column_info_dir)
}
\arguments{
\item{id}{Character. The module ID}

\item{column_info_dir}{Path to directory containing column info files}
}
\value{
List of reactive expressions
}
\description{
Create table picker server
}
